Add confirmation dialog.

+++ frontend/src/app/modules/co-op-configurations/co-op-configurations.component.ts
@@ -19,13 +19,15 @@
 *
 ***************************************************************************/
 
-import { Component, ViewChild, OnInit, ChangeDetectorRef} from '@angular/core';
+import { Component, ViewChild, OnInit, ChangeDetectorRef } from '@angular/core';
 import { MatPaginator, MatPaginatorIntl } from '@angular/material/paginator';
 import { MatTableDataSource } from '@angular/material/table';
 import { MatSort } from '@angular/material/sort';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { CoopConfiguration } from '../../models/co-op-configuration/co-op-configuration';
 import { CoopConfigurationsService } from './services/co-op-configurations.service';
+import { MatDialog } from "@angular/material/dialog";
+import { ConfirmDialogComponent } from '../../modules/confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-co-op-configurations',
@@ -43,7 +45,8 @@ export class CoopConfigurationsComponent implements OnInit {
   showSpinner = false
 
   constructor(private coopConfigurationsService: CoopConfigurationsService,
-    private _snackBar: MatSnackBar) {
+    private _snackBar: MatSnackBar,
+    private dialog: MatDialog) {
     this.paginator = new MatPaginator(new MatPaginatorIntl(), ChangeDetectorRef.prototype);
     this.sort = new MatSort();
     this.coopConfigurations = [];
@@ -67,35 +70,47 @@ export class CoopConfigurationsComponent implements OnInit {
       this.dataSource.data = this.coopConfigurations;
       this.showSpinner = false;
     }).catch(error => {
-        console.error(error);
-        this.openSnackBar(`ERROR: ${error}`);
-        this.showSpinner = false;
+      console.error(error);
+      this.openSnackBar(`ERROR: ${error}`);
+      this.showSpinner = false;
     });
   }
 
   deleteCoopConfiguration(name: string) {
-    this.showSpinner = true;
-    let index = this.coopConfigurations.findIndex(coopConfiguration => {
-      return coopConfiguration.name === name
+    this.dialog
+      .open(ConfirmDialogComponent, {
+        data: {
+          title: 'Delete Co-op Campaign Configuration',
+          message: `Are you sure you want to delete the Co-op Campaign Configuration <strong>${name}</strong>?`
+        }
+      }).afterClosed()
+      .subscribe((confirm: boolean) => {
+        if (confirm) {
+          this.showSpinner = true;
+          let index = this.coopConfigurations.findIndex(coopConfiguration => {
+            return coopConfiguration.name === name
+          });
+          if (index !== -1) {
+            let coopConfiguration: CoopConfiguration = this.coopConfigurations[index];
+            this.coopConfigurationsService.deleteCoopConfiguration(coopConfiguration.name).then((response) => {
+              this.openSnackBar(`The Co-Op Configuration ${coopConfiguration.name} was deleted successfully.`);
+              this.coopConfigurations.splice(index, 1);
+              this.dataSource.data = this.coopConfigurations;
+              this.showSpinner = false;
+            })
+              .catch(error => {
+                console.error(error);
+                this.openSnackBar(`ERROR: ${error}.`);
+                this.showSpinner = false;
+              });
+          } else {
+            console.log(`The Co-Op Configuration was not found.`);
+            this.openSnackBar(`The Co-Op Configuration was not found.`);
+            this.showSpinner = false;
+          }
+
+        }
     });
-    if (index !== -1) {
-      let coopConfiguration : CoopConfiguration = this.coopConfigurations[index];
-      this.coopConfigurationsService.deleteCoopConfiguration(coopConfiguration.name).then((response) => {
-        this.openSnackBar(`The Co-Op Configuration ${coopConfiguration.name} was deleted successfully.`);
-        this.coopConfigurations.splice(index, 1);
-        this.dataSource.data = this.coopConfigurations;
-        this.showSpinner = false;
-      })
-      .catch(error => {
-        console.error(error);
-        this.openSnackBar(`ERROR: ${error}.`);
-        this.showSpinner = false;
-      });
-    } else {
-      console.log(`The Co-Op Configuration was not found.`);
-      this.openSnackBar(`The Co-Op Configuration was not found.`);
-      this.showSpinner = false;
-    }
   }
 
   buildColumns() {
